更新生成器

+++ cmd/gormgen/pkg/template.go
@@ -37,7 +37,7 @@ func NewQueryBuilder() *{{.QueryBuilderName}} {
 	return new({{.QueryBuilderName}})
 }
 
-func (t *{{.StructName}}) Create(db *gorm.DB) (id int32, err error) {
+func (t *{{.StructName}}) Create(db *gorm.DB) (id int64, err error) {
 	if err = db.Create(t).Error; err != nil {
 		return 0, errors.Wrap(err, "create err")
 	}
